add batch file reader as the base class for all file processing
xl_sheet_processor as the base class for all sheet processing logic

Scored2012.py uses the batch file processor to process a set of files

audit_mapper_2012.py uses the xl_sheet_processor for keeping a dict of entries from a sheet and exposes a lookup api

This is probably the most productive checkin that will lead to writing more things on top easier. Older classes could be rewritten using this.